+++ tests/testthat/test_spatstat.R
@@ -1,0 +1,20 @@
+if (requireNamespace("spatstat") && packageVersion("spatstat") >= "2.0.1") {
+  library(spatstat)
+  test_that("Method for converting sfnetwork to linnet works", {
+    roxel_sfn <- as_sfnetwork(roxel) %>% st_transform(3857)
+    # I added suppressWarnings since as.linnet returns a few warning messages
+    # related to the structure of the network (network not connected and
+    # duplicated segments) which are irrelevant here
+    suppressWarnings(roxel_linnet <- as.linnet(roxel_sfn))
+
+    # I need expect_lte since the linnet methods removes the duplicated edges
+    expect_lte(nsegments(roxel_linnet), nrow(roxel))
+  })
+
+  test_that("Method for converting sfnetwork to linnet fails with latlong coordinates", {
+    roxel_sfn <- as_sfnetwork(roxel)
+    expect_error(as.linnet(roxel_sfn))
+  })
+}
+
+